refactor(mobile): reorganize folders
