+++ apps/mobile/src/modules/Home/HomeScreen.tsx
@@ -1,4 +1,4 @@
-import { MapView } from './components';
+import { MapView } from './Map';
 
 export function HomeScreen() {
   return <MapView />;
