--- apps/mobile/src/modules/Home/hooks/index.ts
@@ -1,2 +1,0 @@
-export * from './useCurrentPosition';
-export * from './useInitialRegion';
