+++ apps/mobile/src/modules/Home/index.ts
@@ -1,4 +1,3 @@
 import { HomeScreen } from './HomeScreen';
 
 export default HomeScreen;
-export * from './hooks';
