+++ apps/mobile/src/modules/Onboard/components/index.ts
@@ -1,3 +1,3 @@
-export * from './ExplanationView';
+export * from '../Explanation/ExplanationView';
 export * from './LocationPermissionView';
 export * from './NotificationsPermissionView';
