+++ apps/mobile/src/modules/Onboard/views/Explanation/EarthImage.tsx
