+++ apps/mobile/src/modules/Onboard/views/LocationPermissions/LocationImage.tsx
