+++ apps/mobile/src/modules/Onboard/views/NotificationPermissions/NotificationsImage.tsx
