+++ apps/mobile/src/modules/Onboard/views/index.ts
@@ -1,0 +1,3 @@
+export * from './Explanation/ExplanationView';
+export * from './LocationPermissions/LocationPermissionsView';
+export * from './NotificationPermissions/NotificationPermissionsView';
