+++ apps/mobile/src/shared/hooks/useLocationPermissions.ts
